+++ src/modules/index.ts
@@ -15,4 +15,4 @@ export { PadDirection, runPadCommand } from "./padCommand";
 export { runRemoveBlankLinesCommand } from "./removeBlankLinesCommand";
 export { runRemoveControlCharactersCommand } from "./removeControlCharactersCommand";
 export { runRemoveDuplicatesCommand } from "./removeDuplicatesCommand";
-
+export { runConvertToZalgoCommand, ZalgificationIntensity } from "./convertToZalgoCommand";